Simplified joining of broadcast rooms
The previous code tried to create a broadcast room and setup the alias
if either of these were missing. This introduce a serious race condition
for the discovery room setup for new servers deployment, which resulted
in the failure to run Raiden in our test network.

This commit removes that code that is susceptible to race conditions,
and instead move the responsibility of properly setting up the discovery
room and its alias to the Matrix operator.