Fix matrix test helper
The `NoTLSFederationMonkeyPatchProvider.check_password` function
wasn't a generator function which is required for `defer.inlineCallbacks()`.
It's normally not actually used (we only reach it if the EthAuthProvider
couldn't authenticate a login attempt, which doesn't happen in the tests).

Since we don't perform any deferred calls a simple return statement works
fine.