Fixes for presence checking on first sync
This commit changes when the display name cache is warmed. The goal here
is to not have a cache for usernames that are not whitelisted, since
that is not of interest for the Raidne node and can have a negative
impact for the startup performance, the reason being that the first sync
of the matrix server returns the presence update of every user in a
shared room, and since all Raiden nodes are in the broadcast rooms the
presence a node will receive a presence update for every single node on
the first sync, if a HTTP request is done for each address it can result
in thousands of request to the Matrix server, which has very bad
performance implications.

Additionally this adds a try-except to handle a corner case for a Matrix
federation that has a given userid but the profile of the user is
missing, it seems that this can happen when a server federation is
not properly cleaned (the culprit seems to be a restart of the
federation, were the Matrix servers were cleared, implying in the
removal of the databases, but maybe the federation was not stoped before
the databases were cleared leading to partial synchronization, were some
of the userids were propagated but the profiles were not, so the request
for a given displayname fails with a missing profile error message).
This hypothesis was investigates by @ulope.