Added Matrix sync response queue
The `_sync` has to be done often enough because that is the mechanism
used by the Raiden node to keep it's status as online with the Matrix
server. If a `_sync` returns a huge amount of data that has to be
processed by the node, which may lead to the processing thread taking a
long time to process the server's response, it is possible for the time
in between `_sync` calls can be too large, leading to problems with the
presence checking of the node.

This introduces a message queue, guaranteeing the server responses are
processed in order, since that is necessary for the client state to be
consistent, and that the processing thread will not block the synching
thread.