Accept an invite without having to check whitelist
This was previously done to prevent being invited to many rooms and
being spammed by malicious nodes. However, there are cases where nodes
trying to create rooms for a channel might race and an invite would be
received by one node which did not yet whitelist the inviting node, as
a result the invite would wrongfully be ignored. This change removes
the whitelist check. To prevent spam, we make sure we ignore presence
updates and messages from non-whitelisted nodes.