Pre-fill DisplayNameCache from room membership events
We use the DisplayNameCache to ensure we have a user's displayname
available even in case it's not supplied from the server in certain
cases.

If no cached display name is available it will be fetched from the server.

This improves this slightly by pre-filling the cache from the
`m.room.member` events during first sync (which do contain display names).

This reduces the number of HTTP requests to the matrix server during
startup.