Increase reveal_timeout
We're doing this for now, because the target could be the address
on the lower end of the address order (transport related). Therefore,
the target would be the one to create the room once `ActionInitTarget`
is received. The problem increasing reveal timeout is fixing is that
create-room/invite/accept invite cycle could take time off the lock
where the lock is eventually seen as "in danger zone" after the target
has finished waiting for the initiator to join the room. This results
in the target not asking for the secret and letting the transfer expire.