Matrix: Only create room in the whitelist call
This simplifies the code by only having the room creation at one point,
the whitelist. This works because every node is expected to whitelist
any address it wants to communicate with, so eventually both ends that
want to communicate will do the whitelisting, and the node with the
lower address will eventually create the private room.

There are two scenarios that can happen:
- We have two nodes were both want to communicate with each other, i.e.
when they have a channel open. So, both nodes are expected to do the
whitelisting.
- The communication is initiated assymetrically, i.e. the initiator
sends a transfer to a new target. In this case the message is expected
to arrive off-band, i.e. the node that actually has a channel open with
the target will forward the message with the correct initiator field,
and then both are exptected to have started the whitelisting.