Protect against duplicate room creation
The room creation is currently trigerred by the whitelist method. This
function is used by the blockchain event handler, to whitelist a channel
partner once a channel open event is received. Because of this, it is
possible for the method to be called at the same time by two different
threads, if two channels with the same address pair is open at the same
time. Which would lead to creation of more than one room for the pair
and possible to communication problems.

The lock prevents the concurrent creation of these rooms, preventing the
corner case from above.