Restore reveal timeout
Originally, the tests reveal timeout was increased because it was
suspected that room creation cycle would consume time off the lock's
expiration. However, the reason turned out to be a bug in room
creation that added 10 seconds of latency and
therefore we restore this back to what it was originally.