Do not write the debug logfile in the current directory
Fix #5050

In OSX Catalina the application is launching in a sandbox mode and as
such the current directory would be "/". Naturally raiden is not
allowed to write on the root directory and as such raiden crashes.

This commit attempts to address this by not writing the debug log file in
the current directory but in diffeent directories depending on the OS.

- OSX: ~/Library/Raiden
- Linux: ~/.raiden
- Windows: ~/AppData/Roaming/Raiden