Remove msghash from the default __repr__
To generate the msghash value the message has to be packed into the
binary format, which is very time consuming. This removes the hashing
from the default `__repr__` implementation since the time spent on the
packing was significant (as measured by a sample profiler and later
plotted in a flamegraph).