Added stress test script
This script will orchestrate transfers among nodes in a chain topology,
the goal is to stress the system from the same APIs that an end user
would use, using the real infrastructure.

This script works in rounds, each round will:
- choose the initiator and target pair that have a path in the network.
- plan a set of transfers that are guaranteed to succeed, and that once
  all have complete, the state of the channels is restored to the
  initial position.
- the transfers from the plan are then scheduled in an arbitrary order
- the transfers are executed concurrently and their success is checked.

Besied the above, this script also manages the lifecycle of the Raiden
processes under stress.