Big event fetches done in batches.
The first run can take a bit of time, depending on the number of token
networks registered and when the original registry smart contract was
deployed. This does not improve how long it takes to synchronize with
these old and full registries, but it does make sure that while doing
so, less memory is used, and the work is not lost if the node is
restarted.

This was achieve by using the same approach from the StatelessFilter, in
there the queries are done in batches to avoid timeouts while doing the
request, in the first run this can be used to limit the number of state
changes that are currently held in memory. In order to make sure the
batches of the RaidenService and the underlying filters are alined, the
same constant for the batch size was used.

This does fix one bug, the issue #4498 fixed the problem of not fetching
the events from a newly registered token network on the same batch, but
only for the initialiation and without taking into account the
recoverability. This fixes that bug in a general way, by handling this
corner case in the `synchronize_to_confirmed_block_in_batches` method.
Once that method returns it is known that all events for all the smart
contracts of interest have been handled. And on an event of a crash it
is safe to use the block number from the state machine.

Note that this does not solve the above problem in general, only for
newly registered token networks. The problem in general is a bit harder
since potentially there may be many layers of smart contracts, when
there is a tree of smart contracts that would have to be recursively
followed. The general case would require a form of recursion to handle
all cases.