Bugfix: Don't quit connection attempt when no possible NEW partners
In the code possible new partners means new partners to open channels
with.

We should still try to see if any of our open channels can be funded
by us so that we can reach the `initial_channel_target`.

If it can't the connection stops right below the removed check:

https://github.com/raiden-network/raiden/blob/140c3e7b1b99844c6537f03717233fecff3ae695/raiden/connection_manager.py#L400-L410