Codecov: post to github after coverage is complete
Our jobs submit code coverage separately and codecov.com merges them for
us. Without this change, the coverage informations is posted to github
as soon as it is submitted to codecov. This can irritate developers who
see low coverage for a while until the remaining CI jobs finish and add
their coverage.