Fixed usage of Processed messages
The processed messages were never intended to be a retriable, these
messages offer a stronger guarantee than a Delivered message and are
used for the same purpose, confirming that a given message was received.

The Delivered message was used to confirm messages that have no
dependency on blockchain data, namely SecretRequest and SecretReveal.
While the Processed message were added to have a stronger guarantee, for
messages that require validation against blockchain data. A processed
message is important to fix race conditions, since delivering the
message is not sufficient, it is also necessary for the message to be
sucessfully processed. Since the processing of some messages, namely the
ones with balance proofs, dependo on the on-chain data, the processed
message has to be sent from the state machine.

Here is an example to illustrate the usage of the message:

1. Makes an on-chain deposit to the A-B channel, which is mined at block 10.
2. A sees the confirmation of the deposit first, and sends a message
   that uses the newly available capacity.
3. B receives the LockedTransfers prior to seeing the new Deposit
   on-chain, therefore rejects it.

At this points, A requires a Processed message from B, to know that the
above race condition was fixed. While B doesn't send a Processed message
back to A, the balance proof will be retried.