Unify handling of pfs updates
I'm not sure if this is the right direction to go in, since having the
explicit events is nice. But this is at least consistent instead of
doing some updates as events and others by checking that type of state
change. It's also all in one place and concise.
Testability would be better with the event approach.