APIServer not inheriting from Runnable.
This is the first step to remove the Runnable interface.

The Runnable interface has start/stop methods, in theory one should be
allowed to call `stop` and then `start` on the same instance, which
would restart the service. This however has led to multiple, hard to
debug bugs, because of improper cleanup in `stop`, and partial
initialization in `start`. (E.g. Race conditions among the RestAPI and
RaidenService, invalid re-initialization of the wal and WSGI server,
etc.).

Additionally the Runnable interfaceintroduce cross cutting dependencies.
E.g. the `stop` event must be initialized as part of the `start`
function, otherwise a `stop`/`start` cycle would not work, however, the
`stop_event` is shared across all services, to guarantee that all
services are stoped together, this means a Runnable must not clear the
event, because it does not own it, meaning it is not possible to
implement proper initialization in the `start` methods. (This led to
test failures in the transport layer).