fix race condition on channel deposit
The partner node has to process the channel open before the deposit can
be done. This changes deposit greenlets, so that each will wait for the
channel to be available, and only after the deposit is requested.

It also changes the deposit to be done concurrently with the opening,
once the channel is available the deposit can be done, under some
circumstances it should be faster.