Batch event processing
During initialization of the Raiden node events are restored from the
snapshot and pushed to the transport layer. Each message was queued
individually, which depending on how the greenlets are scheduled by the
gevent's hub may result into multiple unecessary HTTP requests to push
the messags to Matrix. This results in a slow restart of nodes in the
stress test, which may have lots of pending messages in the queue.

This commit fixes the above by pushing the events in batch, which
doesn't rely on the greenlet schduling to reduce the number of HTTP
requests.