Remove monkey patching from `storage/sqlite.py`
This not only broke some filtering code which looked for checksum
encoded addresses, due to the monkeypatching it broke it in a way that
was not detected by the tests because the tests exercised the unpatched
code.

To gain back at least some of the speed improvement, we use
`bytes.hex()` for the networkx graphs when serializing.