Removed Idle measuring
The code did not work properly, the check watcher was scheduled after
the gevent's Hub, which lead the measurement to include the actual
processing done during the loop iteration. Priotizantion of the check
handler didn't work with libev, and gevent's libuv wrapper does not
support check watcher. The last try was to use loop.add_callback which
is a FIFO execution queue, but I couldn't find an easy way of scheduling
the callback only once.

Because there was not reliable way of implementing the measurement this
commit is removing it.