Running message handlers in parallel
Some of the message handler can do IO, e.g. querying the secret registry
to verify if the incoming locked transfer has the secret of the lock
registered or not, these operations can take a while and if executed in
sequence will have a noticiable delay. To remove this delay the
handlers are executed asynchronously in a pool and then the results are
collected.