Decrease transport retry interval
Previously the transport only received a single `retry_interval` parameter.
From that the actual retry interval was calculated as an exponentially
increasing value between `retry_interval` and `retry_interval * 10`.
Additionally the `retries_before_backoff` parameter was set to `5`.

With the previous defaults this means that messages were retried 5 times
every 5 s.

This change splits the interval configuration into an `initial` and a
`max` value.
Also the defaults have been changed to be much less aggressive.
With those values messages will now be retried exponentially increasing
from 5s to the maximum of 60s.