Added async health check.
The initiator has to start a health check to the target on the beginning
of a transfer, this is necessary to processs the secret request message
from the target. The problem is that the health check may trigger the
creation of a room, which adds a measurable amount of latency to the
start of a transfer. This changes the code so that a healthcheck can be
done in the background, allowing this latency to be hidden from the
first transfer.