Test serialized messages for accidental changes
We've recently wasted @kelsos' time by accidentally changing the
signature fore Delivered and Processed messages. This was totally
unnecessary and we should always notice when we change our serialized
messages in any way. This new test should prevent that.

I'm unhappy about the random part, but I also didn't want to change all
messages used in that test file to fixed values. It might be best to
have a set of hypothesis tests to fuzz test the serialization and keep
this new test completely on fixed values. Suggestions welcome!