Temporarily revert to a minimal codecov configuration
In order to turn status reports off completely for now, but still get
coverage graphs.

Due to a bug in the current version of codecov, the status reports
cannot be configured properly and we keep getting failing CI builds
that should pass given our configuration.

This will be reverted when the problem in codecov is fixed.