Detecting unexpected changes to the tracer greenlets
The debugging code instantiates two traces, one installed by the monitor
thread and another used by our debugging code. Each instance installs a
new switch tracer and forwards the value to the previously installed
function. This means that if everything is working as expected both
Tracers will be updated with the same values, and therefore there should
be not discrepancies (There is an edge case, were the monitoring thread
would gain CPU time, *while* the trace callback is being executed, my
assumption is that this is synchronized, otherwise the code for a single
Tracer would already be wrong).

This changes the code as to check if the value differ, and the
asssumption is that if this happened, then there is a problem with the
order of the function cleanup.

Because of the above, just in case the values differ, the code now
raises the exception in the greenlet for Raiden's tracer, instead of
using the value from the monitoring thread, the idea is to maximaly
avoid using the monitoring thread state, because there may be unexpected
side effects on attributes/function calls.