Check exit code when stopping node (stress test)
Without this check, the greenlet will continue past that point and start
the new node even when the old one did not exit cleanly. When starting
the new node, a switch happens right after the Popen, before the new
process is tracked in the nursery. This causes the process to be left
over when the stress test script stops.

To debug and test this error case, I just added an exception during the
raiden shutdown, which show the above behaviour completely reproducibly.

See https://github.com/raiden-network/raiden/issues/5790.

[skip tests]