bugfix: fix download of invalid logs
the previous version of the script was downloading all the logs
available in the server, however, if the scenarios are moved around, it
is possible that the older folders will be dangling. This lead to some
false positives while using the script this week, since the bf5 was
failed on one server, but it was fixed in another.

This should also improve the performance since the files are groupped
together in a tar file, avoid having to reconnect to the server often.

Note: Initially it was tried to download the files individually with
scp, however that was not slower, but also needed more code.

[skip tests]