Reduce the block for the approve lock
The previous code was waiting for the deposit transaction to be mined
and confirmed before releasing the lock, which holds the lock for an
unecessarily long time. This now reduces the critical section just for
the amount of time needed (to ensure there is no other concurrent
approve that would result in errors).