Fixed race-condition
Sending the approve and doing the gas_estimation just after failed. It
looks like the client raced and didn't apply the side-effects of the
pending transaction before doing the gas estimation. The fix was to
introduce a little delay and check if the transaction was added into the
client's pool. The assumption here is that a transaction does not have
to be mined in order for the side-effects to be used in the
estimate_gas.