Removed the invite queue
The queue was only necessary to fix a race condition with whitelisting
and the invite handling, since previously only the invites of
whitelisted addresses were accepted. This is not the case anymore, since
invites are always accepted.

Note: The previous code existed to prevent spamming, the current code
avoids spamming of the internal system by ignoring messages of
non-whitelisted addresses in the handle message. This however is not
sufficient to reduce the load of the /sync calls. For that the code has
to be improved to use the existing filtering APIs.