Allow parallel channel open
Under some corner cases this can break the gas reserve check. This,
however, is not a security issue, and should be a hard to hit corner
case. Since the change to synchronize the open counter and the API
requires a big refactoring to the `new_netting_channel` method, it is
not done here.