fix: set modified while being iterated
The `_processes` set is iterated by `_free_resources`, this function
does operations that result in context switches. At the same time, it
would be possible for a signal for a process that exit to be handled,
and the `subprocess_stopped` would try to remove the element from the
set. This results in the following exception:

    `RuntimeError: Set changed size during iteration`

The fix was to synchronize the access of the `_processes` set with a
reentrant lock.