Checking transaction values.
If `TransactionSlotImplementation` is allocated with invalid values,
sending the transaction will fail but the slot will not be made
available. If that happens a hole will be created in the account's
nonces, and since every new transaction gets the next available nonce
the system will lock until the node is restarted. The asserts were added
to avoid this.