Removed coarse deposit lock
The UserDeposit.deposit_lock was too coarse. Concurrent approves are
handled by the `token_lock`, the only use case for the `depost_lock` was
to serialize calls to the `deposit` function with the goal of avoiding
useless transactions. E.g. if a deposit of 10 and another of 20 are done
at the same time, if the call for the deposit of 20 runs first there is
no need to do the transaction for the value of 10.

Note that concurrent deposits is not a bug from the proxy perspective,
as long as the call itself was valid at the block
`given_block_identifier`.

This commit removes the lock and used a mapping for the currently
in-flight transactions, the other values are check against the
blockchain. This allows multiple deposits to happen concurrently, which
is useful for the scenario player, since the orchestration account sends
a transaction to deposit utility tokens for each Raiden nodes.