Avoid health checking side effects in conn-manager
The healthcheck problems in the bf5 scenarios are caused by
`force_check_address_reachability`, which is only used in the connection
manager. When that is called, the address and its presence become known,
which prevents the room from being created in
`immediate_health_check_for`.

This PR removes all side-effects from
`force_check_address_reachability`, thereby resolving this issue.

Closes https://github.com/raiden-network/raiden/issues/5918