Proper stop of resources for top-level exceptions.
The exception handling code is using `sys.exit` to change the status
code of the process. This immediately stops the process, which means
there the finally block is not run. Because of this the profilling
doesn't get a chance to write the data to disk, effectively discarding
the profilling session. This fixes it by running the cleanup block
before the `sys.exit` calls. The only consequence is that
`write_exception` will not be in the end result, which is not important.