Use PFS with fees in all scenarios
We want to test on mainnet with as few changes as possible to our normal
test runs. Since we have no free services on mainnet, it is easier to
always use a paid PFS.

Contrary to my expectation, this did not require changes in any of the
scenario asserts. This can either be seen as pleasant simplicity (reward
claim amounts are tested in the normal integration tests, after all) or
a gap in the scenario test coverage.