Reduce batch size of the event polling because.
Geth performance for `eth_getLogs` doesn't degrade well with the total
number of events in the blockchain. The testnets response times are
acceptable because the Blockchain does not have that many events,
however, for the Mainnet the number of events is large enough to trigger
timeouts in the client.

Below are the measurements, all o results include time to decode the
response using web3.py (that is why the response time can be higher than
the read timeout):

Geth on Mainnet *without filtering* for Raiden smart contracts.

number of blocks | `eth_getLogs` timing | number of events
0                | 4.715455055236816    | 165
100              | 7.663008451461792    | 14000
200              | 12.996556758880615   | 27559
300              | 15.118395805358887   | 41357
400              | 60.43288326263428    | 56439

Geth on Goerli *without filtering* for Raidne smart contracts:

number of blocks | `eth_getLogs` timing | number of events
0                | 0.025388240814208984 | 2
100              | 0.1299889087677002   | 163
200              | 0.1843571662902832   | 382
300              | 0.22373104095458984  | 533
400              | 0.28818750381469727  | 670

Geth on Mainnet *only for the Raiden smart contracts*.

number of blocks | `eth_getLogs` timing | number of events
0                | 0.051666975021362305 | 0
1000             | 0.9480726718902588   | 28
2000             | 3.340684652328491    | 30
3000             | 4.283600330352783    | 30
4000             | 8.70504117012024     | 30
5000             | 8.81442141532898     | 30
6000             | 8.52564287185669     | 30

The response time increases with the range of blocks in the request. The
reduced batch size required more requests to cover a large number of
blocks, which will lead to slower first sync times. This should be fixed
by #5538, which would result in better for Parity, since its filtering
is much faster.