Fix AssertionError during node startup
The healthcheck uses (via `NotifyingQueue`) an gevent `Event` to wait
for items to be available in the Queue.
If a node already has open channels items will be put into the
healthcheck queue during startup leading to the internal event to
already be set before the first run of the worker.

This triggers gevent/gevent#1540 leading to AssertionErrors to be
printed on stderr.

For now switch the healthcheck worker to a busy loop until the gevent
bug is fixed.