Removed timeout for CLI tests
The timeouts turned to be to strict after the reduction in the block
batch sizes. Because the testchain is across all CLI tests the number of
blocks is unbounded (increases with the number of tests), and because
the test execution order is randomized it is not possible to determine
which tests should have the timeout increased.

This change does not completely remove the timeouts, it just relies on
the `setup_and_call` and `teardown` timeouts setup for the whole suite,
instead of having individual values. The downside is that failures will
take longer to happen, but it should remove the false positives.