Fix problem with pruned blocks.
Consider the following:
- Node A is running, and has channels open with node B
- Node A crashes or is stopped
- Node B closes the channel with node A
- Consider `r` how old an block needs to be to be pruned, and then `r*2`
  blocks passes so that the block with the closed event was pruned a
  long while ago.
- Node A restarts and has to synchronize with the latest changes from
  the blockchain.
- Node A will get its latest confirmed block from the node's state, and
  will start fetching blocks from there up to `h` the current confirmed
  head.
    - If the range above is big enough, not all events are fetched at
      once, since the processing time is slow on the main clients. The
      batch is then split and only a few blocks are fetched every time,
      after every batch a new `c` confirmed block is updated in the node's
      state. If the processing of the event requires additional data,
      the blockchain has to be queried, ideally the block at which the
      event was emitted would be used, bu since `2*r` blocks have passed
      that has been pruned.

The bug in the code was that `c` was used as a fallback, to prevent
errors with pruned data. However, since events are fetched in batches
because of client's performance, `c` itself may have been pruned. The
solution is to use the confirmed head `h` instead.

This assumes that pruning of blocks is not so aggressive that the block
`h` itself will be pruned before the event processing finishes. This
should be fine, as `h` is roughly `unconfirmed head - 5`, and the
pruning should start after at least 128 blocks, which would give a few
minutes for the fetching to run.

This, however, does mean the target has to be continually updated, since
using the block for a long period may wait long enough for it to be
pruned. This is specially a problem for a restart after a long pause,
since the synchronization speed depends on multiple factors, e.g.  the
network latency from the raiden client to the ethereum client, the
latency of the connected ethereum client with the whole network, the
total number of simultaneous applications using that client and how much
overload the client has. For this reason the startup code regulary
fetches the latest block and continue polling until the node is fully
synchronized.

fix #6083