Remove get_channel_participants_from_open_event
The function get_channel_participants_from_open_event relied on fetching
the channel open event from the ethereum node. This, however, is an
operation with linear complexity on both geth and parity, which requires
a lot of CPU from the node and leads to timeout issues. Since this API
can not be safely used in production/mainnet the interface is being
removed, instead the code is relying on the event polling code to be
correct, and assuming that if a NettingChannelState instance is only
created for valid events. This removes the performance drawback from
freshly fecthing the events from the blockchain while maintaining a
level of type safety.

fix #6106